SuiteSparse v5.10.0

+++ GraphBLAS/Source/GB_math.h
@@ -83,11 +83,11 @@
 #define GB_FC32_ne(x,y) ((crealf(x) != crealf(y)) || (cimagf(x) != cimagf(y)))
 #define GB_FC64_ne(x,y) ((creal (x) != creal (y)) || (cimag (x) != cimag (y)))
 
-#define GB_FC32_iseq(x,y) GxB_CMPLXF ((float) GB_FC32_eq (x,y), 0)
-#define GB_FC64_iseq(x,y) GxB_CMPLX  ((float) GB_FC64_eq (x,y), 0)
+#define GB_FC32_iseq(x,y) GxB_CMPLXF ((float)  GB_FC32_eq (x,y), 0)
+#define GB_FC64_iseq(x,y) GxB_CMPLX  ((double) GB_FC64_eq (x,y), 0)
 
-#define GB_FC32_isne(x,y) GxB_CMPLXF ((float) GB_FC32_ne (x,y), 0)
-#define GB_FC64_isne(x,y) GxB_CMPLX  ((float) GB_FC64_ne (x,y), 0)
+#define GB_FC32_isne(x,y) GxB_CMPLXF ((float)  GB_FC32_ne (x,y), 0)
+#define GB_FC64_isne(x,y) GxB_CMPLX  ((double) GB_FC64_ne (x,y), 0)
 
 #define GB_FC32_eq0(x) ((crealf(x) == 0) && (cimagf(x) == 0))
 #define GB_FC64_eq0(x) ((creal (x) == 0) && (cimag (x) == 0))
